feat(auth): add hasAllowedAccess checks in studies

+++ src/controllers/studyController.ts
@@ -25,7 +25,9 @@ import {
 } from '@/common/validation/study-validation.js';
 import { lyricProvider } from '@/core/provider.js';
 import { getDbInstance } from '@/db/index.js';
+import { hasAllowedAccess } from '@/external/authorizationClient.js';
 import { validateRequest } from '@/middleware/requestValidation.js';
+import { extractAccessTokenFromHeader } from '@/middleware/retrieveUserTokenInformation.js';
 import { studyService } from '@/service/studyService.js';
 
 export const getAllStudies = validateRequest(listAllStudies, async (req, res, next) => {
@@ -46,8 +48,14 @@ export const getStudyById = validateRequest(getOrDeleteStudyByID, async (req, re
 	const studyId = req.params.studyId;
 	const db = getDbInstance();
 	const studyRepo = studyService(db);
+	const user = req.user;
+	const token = extractAccessTokenFromHeader(req);
 
 	try {
+		if (!(await hasAllowedAccess(studyId, 'READ', token, user))) {
+			throw new lyricProvider.utils.errors.Forbidden(`You do not have access to study - ${studyId}'`);
+		}
+
 		const results = await studyRepo.getStudyById(studyId);
 		if (!results) {
 			throw new lyricProvider.utils.errors.NotFound(`No Study with ID - ${studyId} found.`);
@@ -63,8 +71,14 @@ export const createNewStudy = validateRequest(createStudy, async (req, res, next
 	const studyData = req.body;
 	const db = getDbInstance();
 	const studyRepo = studyService(db);
+	const user = req.user;
+	const token = extractAccessTokenFromHeader(req);
 
 	try {
+		if (!(await hasAllowedAccess(studyData.studyId, 'WRITE', token, user))) {
+			throw new lyricProvider.utils.errors.Forbidden(`You do not have access to study - ${studyData.studyId}'`);
+		}
+
 		const results = await studyRepo.createStudy(studyData);
 		if (!results) {
 			throw new lyricProvider.utils.errors.BadRequest(`Unable to create study with provided data.`);
@@ -80,8 +94,14 @@ export const deleteStudyById = validateRequest(getOrDeleteStudyByID, async (req,
 	const studyId = req.params.studyId;
 	const db = getDbInstance();
 	const studyRepo = studyService(db);
+	const user = req.user;
+	const token = extractAccessTokenFromHeader(req);
 
 	try {
+		if (!(await hasAllowedAccess(studyId, 'WRITE', token, user))) {
+			throw new lyricProvider.utils.errors.Forbidden(`You do not have access to study - ${studyId}'`);
+		}
+
 		const results = await studyRepo.deleteStudy(studyId);
 
 		if (!results) {
@@ -104,7 +124,14 @@ export const updateStudyById = validateRequest(updateStudy, async (req, res, nex
 	const db = getDbInstance();
 	const studyRepo = studyService(db);
 
+	const user = req.user;
+	const token = extractAccessTokenFromHeader(req);
+
 	try {
+		if (!(await hasAllowedAccess(studyId, 'WRITE', token, user))) {
+			throw new lyricProvider.utils.errors.Forbidden(`You do not have access to study - ${studyId}'`);
+		}
+
 		const results = await studyRepo.updateStudy(studyId, updateData);
 
 		if (!results) {
